requirements.txt
PySimpleGUI
pandas
requests
python-dotenv
jinja2
